redid 1-100

+++ problems/001-100/p007.c
@@ -6,13 +6,13 @@ int main(void) {
 
     int n = 1;
     int size = 0;
-    long* primes = NULL;
+    int* primes = NULL;
     while (size < 10000) {
         free(primes);
         primes = sieve(n, &size);
         n *= 2;
     }
-    printf("Solution: %li\n", primes[10000]);
+    printf("Solution: %d\n", primes[10000]);
 
 
     end_time();
