+++ problems/101-200/p132.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 #define N 1000000000
 
-bool test_prime(long p) {
+bool test_prime(int p) {
     return mod_exp(10, N % (p - 1), p) == 1;
 }
 
@@ -12,7 +12,7 @@ int main(void) {
     start_time();
 
     int size;
-    long* primes = sieve(200000, &size);
+    int* primes = sieve(200000, &size);
 
     long tot = 0;
     int count = 0;
