+++ problems/101-200/p133.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 #define N 1000000000000000000
 
-bool test_prime(long p) {
+bool test_prime(int p) {
     return p >= 7 && mod_exp(10, N % (p - 1), p) == 1;
 }
 
@@ -12,7 +12,7 @@ int main(void) {
     start_time();
 
     int size;
-    long* primes = sieve(100000, &size);
+    int* primes = sieve(100000, &size);
 
     long solution = 0;
     for (int i = 0; i < size; i++) {
